\name{kf_plot}
\alias{kf_plot}
\title{
  plot result of kf_rtn as simply
}
\description{
}
\usage{
kf_plot(rs,xr=1:rs$n,ad=1)
}
\arguments{
  \item{rs}{ result of kf_rtn }
  \item{xr}{ time range }
  \item{ad}{ observation address }
}
\details{
}
\value{
}
\references{
}
\author{
}
\note{
}

\seealso{
}
\examples{
}
\keyword{ kalman filter }
